+++ rust-toolchain.nix
@@ -2,26 +2,23 @@ fenix:
 
 let
   file = ./rust-toolchain.toml;
-  sha256 = "SXRtAuO4IqNOQq+nLbrsDFbVk+3aVA8NNpSZsKlVH/8=";
+  sha256 = "+syqAd2kX8KVa8/U2gz3blIQTTsYYt3U63xBWaGOSc8=";
 in
 {
-  fromFile = { system }: fenix.packages.${system}.fromToolchainFile {
+  fromFile = { buildSystem }: fenix.packages.${buildSystem}.fromToolchainFile {
     inherit file sha256;
   };
 
-  fromTarget = { pkgs, buildPlatform, targetPlatform ? null }:
+  fromTarget = { pkgs, buildSystem, targetSystem }:
     let
-      inherit ((pkgs.lib.importTOML file).toolchain) channel;
-      toolchain = fenix.packages.${buildPlatform};
+      name = (pkgs.lib.importTOML file).toolchain.channel;
+      fenixPackage = fenix.packages.${buildSystem};
+      toolchain = fenixPackage.fromToolchainName { inherit name sha256; };
+      targetToolchain = fenixPackage.targets.${targetSystem}.fromToolchainName { inherit name sha256; };
     in
-    if
-      isNull targetPlatform
-    then
-      fenix.packages.${buildPlatform}.${channel}.toolchain
-    else
-      toolchain.combine [
-        toolchain.${channel}.rustc
-        toolchain.${channel}.cargo
-        toolchain.targets.${targetPlatform}.${channel}.rust-std
-      ];
+    fenixPackage.combine [
+      toolchain.rustc
+      toolchain.cargo
+      targetToolchain.rust-std
+    ];
 }