+++ src/cli.rs
@@ -1,5 +1,12 @@
-use anyhow::Result;
 use clap::{Parser, Subcommand};
+use color_eyre::Result;
+use pimalaya_tui::terminal::{
+    cli::{
+        arg::path_parser,
+        printer::{OutputFmt, Printer},
+    },
+    config::TomlConfig as _,
+};
 use std::path::PathBuf;
 
 #[cfg(feature = "client")]
@@ -22,11 +29,44 @@ pub struct Cli {
 
     /// Override the default configuration file path.
     ///
-    /// The given path is shell-expanded then canonicalized (if
-    /// applicable).
-    #[arg(short, long = "config", global = true)]
-    #[arg(value_name = "PATH", value_parser = config::path_parser)]
-    pub config_path: Option<PathBuf>,
+    /// The given paths are shell-expanded then canonicalized (if
+    /// applicable). If the first path does not point to a valid file,
+    /// the wizard will propose to assist you in the creation of the
+    /// configuration file. Other paths are merged with the first one,
+    /// which allows you to separate your public config from your
+    /// private(s) one(s).
+    #[arg(short, long = "config", global = true, env = "COMODORO_CONFIG")]
+    #[arg(value_name = "PATH", value_parser = path_parser)]
+    pub config_paths: Vec<PathBuf>,
+
+    /// Customize the output format.
+    ///
+    /// The output format determine how to display commands output to
+    /// the terminal.
+    ///
+    /// The possible values are:
+    ///
+    ///  - json: output will be in a form of a JSON-compatible object
+    ///
+    ///  - plain: output will be in a form of either a plain text or
+    ///    table, depending on the command
+    #[arg(long, short, global = true)]
+    #[arg(value_name = "FORMAT", value_enum, default_value_t = Default::default())]
+    pub output: OutputFmt,
+
+    /// Enable logs with spantrace.
+    ///
+    /// This is the same as running the command with `RUST_LOG=debug`
+    /// environment variable.
+    #[arg(long, global = true, conflicts_with = "trace")]
+    pub debug: bool,
+
+    /// Enable verbose logs with backtrace.
+    ///
+    /// This is the same as running the command with `RUST_LOG=trace`
+    /// and `RUST_BACKTRACE=1` environment variables.
+    #[arg(long, global = true, conflicts_with = "debug")]
+    pub trace: bool,
 }
 
 #[derive(Subcommand, Debug)]
@@ -53,16 +93,16 @@ pub enum ComodoroCommand {
 }
 
 impl ComodoroCommand {
-    pub async fn execute(self, config_path: Option<&PathBuf>) -> Result<()> {
+    pub async fn execute(self, printer: &mut impl Printer, config_paths: &[PathBuf]) -> Result<()> {
         match self {
             #[cfg(feature = "client")]
             Self::Timer(cmd) => {
-                let config = TomlConfig::from_some_path_or_default(config_path).await?;
-                cmd.execute(&config).await
+                let config = TomlConfig::from_paths_or_default(config_paths).await?;
+                cmd.execute(printer, &config).await
             }
             #[cfg(feature = "server")]
             Self::Server(cmd) => {
-                let config = TomlConfig::from_some_path_or_default(config_path).await?;
+                let config = TomlConfig::from_paths_or_default(config_paths).await?;
                 cmd.execute(&config).await
             }
             Self::Manual(cmd) => cmd.execute().await,